+++ docs/source/changelog.rst
@@ -4,6 +4,7 @@ Change Log
 
 **Pyro 4.25**
 
+- support for (Linux) abstract namespace AF_UNIX sockets (with a 0-byte at the start of the name)
 - docs: mention that you may have to install serpent manually (most notably with alternative Python implementations)